fix: remove redundant dependency arguments
