+++ src/eravm/mod.rs
@@ -12,7 +12,7 @@ pub mod utils;
 pub use self::r#const::*;
 
 use crate::debug_config::DebugConfig;
-use crate::optimizer::settings::Settings as OptimizerSettings;
+use crate::dependency::Dependency;
 
 use self::context::build::Build;
 use self::context::Context;
@@ -81,7 +81,6 @@ pub fn build_assembly_text(
 ///
 /// Implemented by items which are translated into LLVM IR.
 ///
-#[allow(clippy::upper_case_acronyms)]
 pub trait WriteLLVM<D>
 where
     D: Dependency + Clone,
@@ -112,67 +111,5 @@ where
 {
     fn into_llvm(self, _context: &mut Context<D>) -> anyhow::Result<()> {
         Ok(())
-    }
-}
-
-///
-/// Implemented by items managing project dependencies.
-///
-pub trait Dependency {
-    ///
-    /// Compiles a project dependency.
-    ///
-    fn compile(
-        dependency: Self,
-        path: &str,
-        optimizer_settings: OptimizerSettings,
-        llvm_options: &[String],
-        enable_eravm_extensions: bool,
-        include_metadata_hash: bool,
-        debug_config: Option<DebugConfig>,
-    ) -> anyhow::Result<String>;
-
-    ///
-    /// Resolves a full contract path.
-    ///
-    fn resolve_path(&self, identifier: &str) -> anyhow::Result<String>;
-
-    ///
-    /// Resolves a library address.
-    ///
-    fn resolve_library(&self, path: &str) -> anyhow::Result<String>;
-}
-
-///
-/// The dummy dependency entity.
-///
-#[derive(Debug, Default, Clone)]
-pub struct DummyDependency {}
-
-impl Dependency for DummyDependency {
-    fn compile(
-        _dependency: Self,
-        _path: &str,
-        _optimizer_settings: OptimizerSettings,
-        _llvm_options: &[String],
-        _enable_eravm_extensions: bool,
-        _include_metadata_hash: bool,
-        _debug_config: Option<DebugConfig>,
-    ) -> anyhow::Result<String> {
-        Ok(String::new())
-    }
-
-    ///
-    /// Resolves a full contract path.
-    ///
-    fn resolve_path(&self, _identifier: &str) -> anyhow::Result<String> {
-        Ok(String::new())
-    }
-
-    ///
-    /// Resolves a library address.
-    ///
-    fn resolve_library(&self, _path: &str) -> anyhow::Result<String> {
-        Ok(String::new())
     }
 }
