+++ src/evm/context/mod.rs
@@ -19,9 +19,9 @@ use crate::context::function::declaration::Declaration as FunctionDeclaration;
 use crate::context::function::r#return::Return as FunctionReturn;
 use crate::context::r#loop::Loop;
 use crate::context::IContext;
+use crate::debug_config::DebugConfig;
 use crate::debug_info::DebugInfo;
-use crate::evm::DebugConfig;
-use crate::evm::Dependency;
+use crate::dependency::Dependency;
 use crate::optimizer::Optimizer;
 use crate::target_machine::target::Target;
 use crate::target_machine::TargetMachine;
@@ -67,8 +67,6 @@ where
     /// The manager is used to get information about contracts and their dependencies during
     /// the multi-threaded compilation process.
     dependency_manager: Option<D>,
-    /// Whether to append the metadata hash at the end of bytecode.
-    include_metadata_hash: bool,
     /// The debug info of the current module.
     debug_info: DebugInfo<'ctx>,
     /// The debug configuration telling whether to dump the needed IRs.
@@ -98,7 +96,6 @@ where
         code_type: CodeType,
         optimizer: Optimizer,
         dependency_manager: Option<D>,
-        include_metadata_hash: bool,
         debug_config: Option<DebugConfig>,
     ) -> Self {
         let builder = llvm.create_builder();
@@ -118,7 +115,6 @@ where
             loop_stack: Vec::with_capacity(Self::LOOP_STACK_INITIAL_CAPACITY),
 
             dependency_manager,
-            include_metadata_hash,
             debug_info,
             debug_config,
 
@@ -202,22 +198,13 @@ where
     }
 
     ///
-    /// Compiles a contract dependency, if the dependency manager is set.
+    /// Get the contract dependency data.
     ///
-    pub fn compile_dependency(&mut self, name: &str) -> anyhow::Result<String> {
+    pub fn get_dependency_data(&self, path: &str) -> anyhow::Result<String> {
         self.dependency_manager
-            .to_owned()
+            .as_ref()
             .ok_or_else(|| anyhow::anyhow!("The dependency manager is unset"))
-            .and_then(|manager| {
-                Dependency::compile(
-                    manager,
-                    name,
-                    self.optimizer.settings().to_owned(),
-                    self.llvm_options.as_slice(),
-                    self.include_metadata_hash,
-                    self.debug_config.clone(),
-                )
-            })
+            .and_then(|manager| Dependency::get(manager, path))
     }
 
     ///
